third_party/blink/web_tests/platform/generic/editing/input/scroll-viewport-page-up-down-expected.txt

Page up/down (option+page up/down on Mac) should move the move cursor and scroll the content in contenteditable elements. This sample covers scroll position test of a) iframe element containing contenteditable body and b) content editable div element. Even though the cursor will move exactly to the same location on all platforms (covered by test option-page-up-down.html), please note that Mac will scroll the visible area by placing the cursor position in the middle. All other platforms will scroll by keeping the cursor aligned with the top edge of the visible area.

line 0

line 1

line 2

line 3

line 4

line 5

line 6

line 7

line 8

line 9

iframe PASS
div PASS
